Actualización última clase

+++ Controllers/ProductoController.cs
@@ -17,6 +17,11 @@ namespace YerovyInventoryService.Controllers
 
         public IActionResult Index()
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             var productos = _context.Productos
                 .Include(p => p.Categoria)
                 .ToList();
@@ -25,14 +30,39 @@ namespace YerovyInventoryService.Controllers
         }
 
         //Guardar producto
+
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            ViewBag.Categorias = _context.Categorias.ToList();
             return View();
         }
 
         [HttpPost]
-        public IActionResult Create(Producto producto)
+        public IActionResult Create(Producto producto, IFormFile imagen)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (imagen != null)
+            {
+                var ruta = Path.Combine(Directory.GetCurrentDirectory(),
+                    "wwwroot/images", imagen.FileName);
+
+                using (var stream = new FileStream(ruta, FileMode.Create))
+                {
+                    imagen.CopyTo(stream);
+                }
+
+                producto.ImagenUrl = "/images/" + imagen.FileName;
+            }
+
             _context.Productos.Add(producto);
             _context.SaveChanges();
 
@@ -42,15 +72,74 @@ namespace YerovyInventoryService.Controllers
         //Formulario editar
         public IActionResult Edit(int id)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             var producto = _context.Productos.Find(id);
             ViewBag.Categorias = _context.Categorias.ToList();
 
             return View(producto);
         }
 
+        [HttpPost]
+        public IActionResult Edit(Producto producto, IFormFile imagen)
+        {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            var productoBD = _context.Productos.Find(producto.Id);
+            if (productoBD == null)
+                return NotFound();
+            //Actualizar datos normales
+            productoBD.Nombre = producto.Nombre;
+            productoBD.Precio = producto.Precio;
+            productoBD.Stock = producto.Stock;
+            productoBD.CategoriaId = producto.CategoriaId;
+
+            //Si sube nueva imagne
+            if (imagen != null)
+            {
+                var carpeta = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/images");
+
+                if (!Directory.Exists(carpeta))
+                {
+                    Directory.CreateDirectory(carpeta);
+                }
+
+                var ruta = Path.Combine(carpeta, imagen.FileName);
+
+                using (var stream = new FileStream(ruta, FileMode.Create))
+                {
+                    imagen.CopyTo(stream);
+                }
+                productoBD.ImagenUrl = "/mages/" + imagen.FileName;
+            }
+
+            _context.SaveChanges();
+
+            return RedirectToAction("Index");
+        }
+
         //Eliminar producto
         public IActionResult Delete(int id)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            var rol = HttpContext.Session.GetString("Rol");
+
+            //SOLO ADMIN PUEDE ELIMINAR
+            if (rol != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             var producto = _context.Productos.Find(id);
 
             _context.Productos.Remove(producto);
